integrity: move cli/integrity to lib/integrity
Part of #757
