+++ cli/cluster/cmd/common.go
@@ -10,7 +10,7 @@ import (
 	"github.com/tarantool/go-tarantool"
 	"github.com/tarantool/tt/cli/cluster"
 	"github.com/tarantool/tt/cli/connect"
-	"github.com/tarantool/tt/cli/integrity"
+	"github.com/tarantool/tt/lib/integrity"
 )
 
 // printRawClusterConfig prints a raw cluster configuration or an instance
