+++ cli/cluster/collectorfactory_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/tarantool/go-tarantool"
 	"github.com/tarantool/tt/cli/cluster"
-	"github.com/tarantool/tt/cli/integrity"
+	"github.com/tarantool/tt/lib/integrity"
 )
 
 func TestDataCollectorFactory(t *testing.T) {
