+++ cli/cmd/pack.go
@@ -7,9 +7,9 @@ import (
 	"github.com/apex/log"
 	"github.com/spf13/cobra"
 	"github.com/tarantool/tt/cli/cmdcontext"
-	"github.com/tarantool/tt/cli/integrity"
 	"github.com/tarantool/tt/cli/modules"
 	"github.com/tarantool/tt/cli/pack"
+	"github.com/tarantool/tt/lib/integrity"
 )
 
 // packCtx contains information for tt pack command.
