+++ cli/pack/common.go
@@ -15,9 +15,9 @@ import (
 	"github.com/tarantool/tt/cli/cmdcontext"
 	"github.com/tarantool/tt/cli/config"
 	"github.com/tarantool/tt/cli/configure"
-	"github.com/tarantool/tt/cli/integrity"
 	"github.com/tarantool/tt/cli/running"
 	"github.com/tarantool/tt/cli/util"
+	"github.com/tarantool/tt/lib/integrity"
 	lua "github.com/yuin/gopher-lua"
 	"gopkg.in/yaml.v2"
 )
