+++ cli/running/script_instance_test.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/tarantool/tt/cli/integrity"
 	"github.com/tarantool/tt/cli/ttlog"
+	"github.com/tarantool/tt/lib/integrity"
 )
 
 const (
